requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.11.2
bson==0.5.10
cachelib==0.10.2
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
dominate==2.7.0
filelock==3.9.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
Pillow==9.4.0
pipenv==2023.2.4
platformdirs==3.0.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-http-client==3.3.7
pytz==2022.7.1
requests==2.28.2
sendgrid==6.9.7
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.3
starkbank-ecdsa==2.2.0
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.19.0
virtualenv-clone==0.5.7
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1
